requirements.txt
pkgconfig
pytezos==3.3.0
requests==2.25.1
python-dotenv==0.19.2
simplejson
ply
cattrs-extras==0.1.1
mnemonic
cached-property==1.5.2
pysodium==0.7.7
tabulate==0.8.10
# adal
